BaseConverters test done and passed

+++ DocumentModel.BaseTypes/Converters/EMUJsonConverter.cs
@@ -83,7 +83,6 @@ public class EMUJsonConverter : JsonConverter<EMU>
   /// </remarks>
   public override void Write(Utf8JsonWriter writer, EMU value, JsonSerializerOptions options)
   {
-    // Write as numeric string without unit suffix
-    writer.WriteStringValue(((Int64)value).ToString(CultureInfo.InvariantCulture));
+    writer.WriteStringValue(value.ToString(CultureInfo.InvariantCulture));
   }
 }
