+++ DocumentModel.BaseTypes/Converters/EPSJsonConverter.cs
@@ -101,7 +101,6 @@ public class EPSJsonConverter : JsonConverter<EPS>
   /// </remarks>
   public override void Write(Utf8JsonWriter writer, EPS value, JsonSerializerOptions options)
   {
-    // Write as numeric string without unit suffix
-    writer.WriteStringValue(((Int64)value).ToString(CultureInfo.InvariantCulture));
+    writer.WriteStringValue(value.ToString(CultureInfo.InvariantCulture));
   }
 }
