+++ DocumentModel.BaseTypes/Converters/HPSJsonConverter.cs
@@ -84,7 +84,6 @@ public class HPSJsonConverter : JsonConverter<HPS>
   /// </remarks>
   public override void Write(Utf8JsonWriter writer, HPS value, JsonSerializerOptions options)
   {
-    // Write as numeric string without unit suffix
-    writer.WriteStringValue(((Int64)value).ToString(CultureInfo.InvariantCulture));
+    writer.WriteStringValue(value.ToString(CultureInfo.InvariantCulture));
   }
 }
